backend/requirements.txt
fastapi
uvicorn[standard]

pydantic>=2.4.1,<2.12
pydantic-settings==2.5.2

httpx
PyYAML
